Split reminders.ts into sub-modules to stay within 1000-line limit
functions/_lib/services/reminders.ts is now a thin orchestrator.
Logic for each reminder type lives in its own file under reminders/:
- shared.ts       — batch helpers and invite-filter utilities
- invite-reminders.ts        — attendee + speaker invite reminders
- cospeaker-reminders.ts     — co-speaker invite reminders
- presentation-reminders.ts  — presentation upload reminders
- confirmation-reminders.ts  — registration confirmation reminders + cancellations
